# trackClickHandler optimizations