Called 100 times in a loop for 6.7k tracks in a playlist.
```
Firefox
master
default: 6787ms - timer ended

replace .each
default: 5503ms - timer ended

1 handler
default: 2402ms - timer ended

all
default: 2272ms - timer ended

Brave:
master
default: 8478.673828125ms

all
default: 3807.89013671875ms
```
~Remaining time is still way too high~, but I'll give myself a brief pause from AP. I think remaining seconds are mainly caused by getting the \<li> elements and that could be avoided with a clever idea.
Edit: sorry, I was off by one decimal point. It's 0.06/0.02 second per call. OK :)